+++ Source/SwiftLintFramework/Rules/ValidDocsRule.swift
@@ -39,14 +39,18 @@ func superfluousOrMissingThrowsDocumentation(declaration: String, comment: Strin
     return declaration.containsString(" throws ") == !comment.containsString("- throws:")
 }
 
-func delcarationReturns(declaration: String, kind: SwiftDeclarationKind) -> Bool {
-    if SwiftDeclarationKind.variableKinds().contains(kind) { return true }
+func delcarationReturns(declaration: String, kind: SwiftDeclarationKind? = nil) -> Bool {
+    if let kind = kind where SwiftDeclarationKind.variableKinds().contains(kind) {
+        return true
+    }
+    guard let outsideBracesMatch = regex("(?:\\)(\\s*->\\s*)?[^()->]*(\\(.*\\))*\\s*\\{)")
+        .matchesInString(declaration, options: [],
+            range: NSRange(location: 0, length: declaration.characters.count)).first else {
+                return false
+    }
 
-    let outsideBraces = NSMutableString(string: declaration)
-    regex("(\\s*->\\s*)[^(]*\\)").replaceMatchesInString(outsideBraces, options: [],
-        range: NSRange(location: 0, length: outsideBraces.length), withTemplate: "")
-
-    return outsideBraces.containsString("->")
+    return NSString(string: declaration).substringWithRange(outsideBracesMatch.range)
+        .containsString("->")
 }
 
 func commentHasBatchedParameters(comment: String) -> Bool {
@@ -59,11 +63,7 @@ func commentReturns(comment: String) -> Bool {
 }
 
 func missingReturnDocumentation(declaration: String, comment: String) -> Bool {
-    let outsideBraces = NSMutableString(string: declaration)
-    regex("(\\s*->\\s*)[^(]*\\)").replaceMatchesInString(outsideBraces, options: [],
-        range: NSRange(location: 0, length: outsideBraces.length), withTemplate: "")
-
-    return outsideBraces.containsString("->") && !commentReturns(comment)
+    return delcarationReturns(declaration) && !commentReturns(comment)
 }
 
 func superfluousReturnDocumentation(declaration: String, comment: String,
@@ -142,6 +142,15 @@ public struct ValidDocsRule: ConfigProviderRule {
             "\npublic func no(param: (Void -> Void)?, param2: String->Void) {}",
             "/// docs👨‍👩‍👧‍👧\n/// - returns: false\npublic func no() -> Bool { return false }",
             "/// docs\n/// - returns: tuple\npublic func no() -> (Int, Int) {return (1, 2)}",
+            "/// docs\n/// - returns: closure\npublic func no() -> (Void->Void) {}",
+            "/// docs\n/// - parameter param: this is void" +
+            "\n/// - parameter param2: this is void too" +
+            "\nfunc no(param: (Void) -> Void, onError param2: ((NSError) -> Void)? = nil) {}",
+            "/// docs\n/// - parameter param: this is a void closure" +
+            "\n/// - parameter param2: this is a void closure too" +
+            "\n/// - parameter param3: this is a void closure too" +
+            "\nfunc a(param: () -> Void, param2: (parameter: Int) -> Void, " +
+            "param3: (parameter: Int) -> Void) {}",
         ],
         triggeringExamples: [
             "/// docs\npublic ↓func a(param: Void) {}\n",
@@ -163,6 +172,10 @@ public struct ValidDocsRule: ConfigProviderRule {
             "/// docs\n/// - parameter param: this is void" +
             "\n///- parameter param2: this is void too\n/// - returns: false" +
             "\npublic ↓func no(param: (Void -> Void)?, param2: String->Void) {}",
+            "/// docs\npublic func no() -> (Int, Int) {return (1, 2)}",
+            "/// docs\n/// - parameter param: this is void" +
+            "\n///- parameter param2: this is void too\n///- returns: closure" +
+            "\nfunc no(param: (Void) -> Void, onError param2: ((NSError) -> Void)? = nil) {}",
         ]
     )